feat: :sparkles: Pdf
Endpoint para generar el pdf. aqui acaba la v1 de la API

+++ app/Controllers/UsersController.php
@@ -78,7 +78,7 @@ class UsersController extends BaseController
             $input = $this->getRequestInput($this->request);
             $photoFile = $this->request->getFile('photo');
 
-            if ($photoFile->isValid() && !$photoFile->hasMoved()) {
+            if ($photoFile !== null && $photoFile->isValid() && !$photoFile->hasMoved()) {
                 if (!empty($user['photo'])) {
                     unlink(ROOTPATH . 'public/' . $user['photo']);
                 }
@@ -168,35 +168,142 @@ class UsersController extends BaseController
      *
      * @return \CodeIgniter\HTTP\Response
      */
-    public function generateUserListPdf(): Response
+    public function generateUserListPdf(): ResponseInterface
     {
-        $mpdf = new \Mpdf\Mpdf();
+        try {
+            $pdf = new \TCPDF();
+            $pdf->setPrintHeader(false);
+            $pdf->setPrintFooter(false);
+            $pdf->AddPage();
 
-        $userModel = new UserModel();
-        $users = $userModel->findAll();
+            $model = new UserModel();
+            $users = $model->findAll();
 
-        $html = '<h1>Listado de Usuarios</h1>';
-        foreach ($users as $user) {
-            $html .= '<img src="' . site_url('public/' . $user['photo']) . '" width="100" height="100"><br>';
-            $html .= 'Nombre: ' . $user['name'] . ' ' . $user['last_name'] . '<br>';
-            $html .= 'Celular: ' . $user['phone'] . '<br>';
-            $html .= 'Correo electrónico: ' . $user['email'] . '<br>';
-            $html .= 'Tipo de Usuario: ' . $user['type_user'] . '<br><br>';
+            if (!(count($users))) {
+                return $this->getResponse([
+                    'message' => 'No users found',
+                ], ResponseInterface::HTTP_NOT_FOUND);
+            }
+
+            $html = '<h1>List of Users</h1>';
+            $html .= '<table>';
+            $html .= '<tr><th>Name</th><th>Email</th></tr>';
+
+            foreach ($users as $user) {
+                $html .= '<tr>';
+                $html .= '<td>' . $user['name'] . ' ' . $user['last_name'] . '</td>';
+                $html .= '<td>' . $user['email'] . '</td>';
+                $html .= '</tr>';
+            }
+
+            $html .= '</table>';
+
+            $pdf->writeHTML($html, true, false, true, false, '');
+
+            $pdfFileName = 'user_list_' . date('YmdHis') . '.pdf';
+
+            $pdfFilePath = WRITEPATH . 'pdfs/' . $pdfFileName;
+            $pdf->Output($pdfFilePath, 'F');
+
+            $pdfUrl = base_url('pdfs/' . $pdfFileName);
+
+            return $this->getResponse([
+                'message' => 'PDF generated successfully',
+                'pdf_url' => $pdfUrl,
+            ], ResponseInterface::HTTP_OK);
+        } catch (\Exception $exception) {
+            return $this->getResponse([
+                'error' => $exception->getMessage()
+            ], ResponseInterface::HTTP_INTERNAL_SERVER_ERROR);
         }
+    }
 
-        $mpdf->WriteHTML($html);
+    /**
+     * Retrieve the authenticated user.
+     *
+     * @return ResponseInterface The HTTP response containing user details.
+     */
+    public function profile(): ResponseInterface
+    {
+        $authenticationHeader = $this->request->getServer('HTTP_AUTHORIZATION');
 
-        $pdfFileName = 'user_list.pdf';
-        $pdfFilePath = WRITEPATH . 'uploads/' . $pdfFileName;
-        $mpdf->Output($pdfFilePath, 'F');
+        try {
+            helper('jwt');
+            $encodedToken = getJWTFromRequest($authenticationHeader);
+            $decodedToken = validateJWTFromRequest($encodedToken);
 
-        $pdfUrl = base_url('uploads/' . $pdfFileName);
+            $model = new UserModel();
+            $user = $model->findUserByPhone($decodedToken->phone);
 
-        $response = [
-            'message' => 'PDF generate successfully',
-            'pdf_url' => $pdfUrl,
-        ];
+            if ($user) {
+                unset($user['password']);
 
-        return $this->response->setJSON($response);
+                $lastLogin = $model->getLastLogin($user['id']);
+                $user['last_login'] = $lastLogin ? $lastLogin['login_time'] : null;
+
+                return $this->getResponse([
+                    'message' => 'User retrieved successfully',
+                    'user' => $user
+                ]);
+            } else {
+                return $this->getResponse(['error' => 'Record not found'], ResponseInterface::HTTP_NOT_FOUND);
+            }
+        } catch (\Exception $exception) {
+            return $this->getResponse([
+                'error' => $exception->getMessage()
+            ], ResponseInterface::HTTP_UNAUTHORIZED);
+        }
+    }
+
+    /**
+     *  Update authenticated user data.
+     *
+     * @return ResponseInterface The HTTP response containing user details.
+     */
+    public function updateProfile(): ResponseInterface
+    {
+        $authenticationHeader = $this->request->getServer('HTTP_AUTHORIZATION');
+
+        try {
+            helper('jwt');
+            $encodedToken = getJWTFromRequest($authenticationHeader);
+            $decodedToken = validateJWTFromRequest($encodedToken);
+
+            $model = new UserModel();
+            $user = $model->findUserByPhone($decodedToken->phone);
+
+            if ($user) {
+                $input = $this->getRequestInput($this->request);
+                $photoFile = $this->request->getFile('photo');
+
+                if ($photoFile !== null && $photoFile->isValid() && !$photoFile->hasMoved()) {
+                    $newName = $photoFile->getRandomName();
+                    $photoFile->move(ROOTPATH . 'public/uploads', $newName);
+                    $photoPath = 'uploads/' . $newName;
+                    $input['photo'] = $photoPath;
+                } else {
+                    unset($input['photo']);
+                }
+
+                $model->update($user['id'], $input);
+
+                $updatedUser = $model->findUserByPhone($user['phone']);
+
+                $lastLogin = $model->getLastLogin($user['id']);
+                $updatedUser['last_login'] = $lastLogin ? $lastLogin['login_time'] : null;
+                unset($updatedUser['password']);
+
+                return $this->getResponse([
+                    'message' => 'User profile updated successfully',
+                    'user' => $updatedUser
+                ]);
+            } else {
+                return $this->getResponse(['error' => 'User not found'], ResponseInterface::HTTP_NOT_FOUND);
+            }
+        } catch (\Exception $exception) {
+            return $this->getResponse([
+                'error' => $exception->getMessage()
+            ], ResponseInterface::HTTP_UNAUTHORIZED);
+        }
     }
 }
